Client: Notify about exposure changes when popup is repositioned
QXdgSurface::Popup::setWindowPosition() changes the exposed state, but
it doesn't call QWindowSystemInterface::handleExposeEvent() to notify
the QGuiApplication layer about it.

For example, this can result in the QMenu unexpectedly repainting with
the new size even though QtWayland still waits for the reposition
request to be completed.

xdg_popup_repositioned doesn't need to call updateExposure() because
it is going to be followed by a configure event and
QWaylandXdgSurface::applyConfigure() already updates the exposed state.

Pick-to: 6.9
Change-Id: Ia0435e2cd1a0d39e0edac61beb4c0d761314c1d6
Reviewed-by: David Edmundson <davidedmundson@kde.org>