qgraphicsitem_cast: replace runtime with compile-time check
Specialize the cast template for QGraphicsItem, where we can skip the
runtime comparisons of the type.

In addition, replace the runtime check for QGraphicsItem::Type with a
compile time check in the main template.

We need to do both so that we correctly cast up to QGraphicsItem for
custom item types that don't provide their own Type alias.

Pick-to: 6.9 6.8
Change-Id: Ic1bff3404fe890747865ce1349cddbcfebb3b77b
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>