macOS: Remove drawRect and use plain CALayer for raster windows
The drawRect callback was used when we were supporting both surface-
backed and layer-backed drawing, but nowadays we only support layer-
backed drawing.

However we couldn't just remove the implementation, as the default
backing layer of an NSView is an NSViewBackingLayer, which in its
override of [CALayer display] checks whether the view implements
drawRect, and if not bails out, which meant we never reached our
displayLayer callback from [CALayer display].

But, we don't need NSViewBackingLayer, as we handle layer updates
and draw callbacks in a more generic way, to support CAMetalLayer,
so instead we now use a plain CALayer as our backing layer for
raster windows.

Change-Id: I5b878ff51ca92dafcb80bf5a686ced13088cfa28
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>