QUrl: use some QStringView internally a bit more
Implemented for the internal QUrlPrivate::setXxxx functions that don't
call setError() (that is, all except setAuthority() and setHost())
because setError() takes the full input to store for the error string.

This temporarily means we won't copy the user's QString when no
modifications are made to store in the PrettyDecoded mode. That will be
fixed in the next commit.

Change-Id: I984e855c964c23c93c4dfffd1a703ca00ed2090e
Reviewed-by: David Faure <david.faure@kdab.com>