QAbstractSlider: remove duplicate update() calls from setOrientation()
The sliderChange() base implementation already calls update(), and the
only reason we left the update() call in when we added the missing
sliderChange() call to setOrientation() was to play it extra safe in
the face of misbehaving overrides.

For 6.10, make sure to not penalize correct programs anymore.

[ChangeLog][Important Behavior Changes][QtWidgets][QAbstractSlider]
setOrientation() now relies on sliderChange() to call update(). This
was already the case for setRange(), setSingleStep() and
setPageStep(), and setOrientation() was only handled specially because
of QTBUG-135597.

Task-number: QTBUG-135597
Change-Id: I71ec5caf61bc656c2ef3421b4de696eac8623e8e
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>