rhi: Remove runCleanup function
The rest of the framework uses cleanup callbacks in a way that implies
that manually triggering the callbacks will lead to problems:
components such as QQuickWidget register cleanup callbacks to get
notified about the destruction of the QRhi. These callbacks should
not be invoked at arbitrary times.

There is no usage of this function anywhere in Qt. Remove it.

Slightly extend the autotest by exercising the keyed registration
functions as well.

Change-Id: I88f1a1e9bc5a642b8e8b6238fe198f123bd55978
Reviewed-by: Andy Nichols <andy.nichols@qt.io>