QLoggingRegistry: remove use of QDir and merge two strings
We don't need to apply QDir::absolutePath() on the diretory that
QLibraryInfo::path() returns, because that is already an absolute path
(and if it weren't, it wouldn't make a difference anyway). That means we
are only doing path concatenation here. For code this low-level, it's
better to stick to QString.

This commit also rejiggles the string variables so we avoid one of the
memory allocations when calling QStandardPaths.

Pick-to: 6.9
Change-Id: Ic3405877155a15ac4ac6fffdc92463ae0bff0bd9
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>